requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
certifi==2025.1.31
distlib==0.3.9
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
environ==1.0
filelock==3.17.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ollama==0.4.7
platformdirs==4.3.6
psycopg2-binary==2.9.10
pydantic==2.11.1
pydantic_core==2.33.0
PyJWT==2.9.0
PyMuPDF==1.25.5
python-decouple==3.8
sniffio==1.3.1
sqlparse==0.5.3
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.1
virtualenv==20.29.3
